Explicit uint16_t casts in the base32768 reverse-table init
MSVC flagged the size_t -> uint16_t narrowing as C4267 on the Windows
wheel build; the values fit (< 32768 and < 128), so the casts state the
guarantee. Line 33 casts the whole OR: integer promotion would
otherwise hand MSVC an int and a C4244 instead.

+++ src/radixly/base32768/base32768.c
@@ -26,11 +26,11 @@ radixly_base32768_exec(PyObject *Py_UNUSED(module))
     }
 
     for (size_t i = 0; i < RADIXLY_ARRAY_SIZE(RADIXLY_B32768_FWD15); i++) {
-        REV[RADIXLY_B32768_FWD15[i]] = i;
+        REV[RADIXLY_B32768_FWD15[i]] = (uint16_t)i;
     }
 
     for (size_t i = 0; i < RADIXLY_ARRAY_SIZE(RADIXLY_B32768_FWD7); i++) {
-        REV[RADIXLY_B32768_FWD7[i]] = REV_7BIT_FLAG | i;
+        REV[RADIXLY_B32768_FWD7[i]] = (uint16_t)(REV_7BIT_FLAG | i);
     }
     return 0;
 }
